+++ src/app/components/reservation/reservation.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ReservationService }from 'src/app/services/reservation.service';
 import { Reservation } from 'src/app/models/reservation';
+import { HttpErrorResponse } from '@angular/common/http';
 
 @Component({
   selector: 'app-reservation',
@@ -21,9 +22,37 @@ export class ReservationComponent implements OnInit {
   }
 
   deleteReservation(reservationId: number) {
-    this.reservationService.deleteReservation(reservationId).subscribe(() => {
-      this.reservations = this.reservations.filter(item => item.id !== reservationId);
-      console.log('Reservation deleted successfully!');
-    })
+    if (confirm('Do you want to delete this reservations?')) {
+      this.reservationService.deleteReservation(reservationId).subscribe(() => {
+        this.reservations = this.reservations.filter(item => item.id !== reservationId);
+        console.log('Reservation deleted successfully!');
+      })
+    }
+  }
+
+  public searchReservations(key: string): void {
+    console.log(key);
+    const results: Reservation[] = [];
+    for (const reservation of this.reservations) {
+      if (reservation.conferenceRoom.name.toLowerCase().indexOf(key.toLowerCase()) !== -1
+      || reservation.startDateTime.toLowerCase().indexOf(key.toLowerCase()) !== -1) {
+        results.push(reservation);
+      }
+    }
+    this.reservations = results;
+    if (results.length === 0 || !key) {
+      this.getReservations();
+    }
+  }
+
+  public getReservations(): void {
+    this.reservationService.getReservations().subscribe((data: Reservation[]) => {
+      this.reservations = data;
+      console.log(this.reservations);
+      },
+      (error: HttpErrorResponse) => {
+        alert(error.message);
+      }
+    );
   }
 }
